modify pkg name
