+++ pkg/cli/upgradeassistant/internal/repository/models/cluster.go
@@ -14,20 +14,16 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package _80
+package models
 
 import (
 	"go.mongodb.org/mongo-driver/bson/primitive"
 )
 
-type ProjectClusterRelation struct {
-	ID          primitive.ObjectID `json:"id,omitempty"              bson:"_id,omitempty"`
-	ProjectName string             `json:"project_name"              bson:"project_name"`
-	ClusterID   string             `json:"cluster_id"                bson:"cluster_id"`
-	CreatedAt   int64              `json:"createdAt"                 bson:"createdAt"`
-	CreatedBy   string             `json:"createdBy"                 bson:"createdBy"`
+type K8SCluster struct {
+	ID primitive.ObjectID `json:"id,omitempty"              bson:"_id,omitempty"`
 }
 
-func (ProjectClusterRelation) TableName() string {
-	return "project_cluster_relation"
+func (K8SCluster) TableName() string {
+	return "k8s_cluster"
 }
